+++ packages/react/src/components/DataTable/src/components/DataTableDropdown/index.ts
@@ -1,0 +1,2 @@
+export { DataTableDropdown } from "./DataTableDropdown";
+export type { DataTableDropdownProps } from "./dataTableDropdown.types";