+++ packages/react/src/components/DataTable/src/components/DataTableDropdownDivider/index.ts
@@ -1,0 +1,1 @@
+export { DataTableDropdownDivider } from "./DataTableDropdownDivider";